fix simd and perf issues, add a tool for puppeteer perf data
Add Puppeteer code for perfing debug and experimental_SIMD builds of CanvasKit against all .skp's

+++ tools/perf-canvaskit-puppeteer/perf-canvaskit-with-puppeteer.js
@@ -48,6 +48,16 @@ const opts = [
     description: 'The perf file to write. Defaults to perf.json',
   },
   {
+    name: 'chromium_executable_path',
+    typeLabel: '{underline file}',
+    description: 'The chromium executable to be used by puppeteer to run tests',
+  },
+  {
+    name: 'overwrite_existing_output_property',
+    typeLabel: String,
+    description: 'Overwrites a json property in an existing output file.',
+  },
+  {
     name: 'use_gpu',
     description: 'Whether we should run in non-headless mode with GPU.',
     type: Boolean,
@@ -186,6 +196,7 @@ async function driveBrowser() {
       // Chrome instance will be used instead of puppeteer spinning up a new one.
       '--disable-frame-rate-limit',
       '--disable-gpu-vsync',
+      '--enable-features=WebAssemblySimd'
   ];
   if (options.use_gpu) {
     browser_args.push('--ignore-gpu-blacklist');
@@ -194,7 +205,11 @@ async function driveBrowser() {
   }
   console.log("Running with headless: " + headless + " args: " + browser_args);
   try {
-    browser = await puppeteer.launch({headless: headless, args: browser_args});
+    browser = await puppeteer.launch({
+      headless: headless,
+      args: browser_args,
+      executablePath: options.chromium_executable_path
+    });
     page = await browser.newPage();
     await page.setViewport(viewPort);
   } catch (e) {
@@ -252,7 +267,19 @@ async function driveBrowser() {
     } else {
       const perfResults = await page.evaluate('window._perfData');
       console.debug('Perf results: ', perfResults);
-      fs.writeFileSync(options.output, JSON.stringify(perfResults));
+
+      if (options.overwrite_existing_output_property) {
+        const existing_output_file_contents = fs.readFileSync(options.output, 'utf8');
+        let existing_dataset = {};
+        try {
+          existing_dataset = JSON.parse(existing_output_file_contents);
+        } catch (e) {}
+
+        existing_dataset[options.overwrite_existing_output_property] = perfResults;
+        fs.writeFileSync(options.output, JSON.stringify(existing_dataset));
+      } else {
+        fs.writeFileSync(options.output, JSON.stringify(perfResults));
+      }
     }
 
   } catch(e) {
